weight to float

+++ ecommerce/apps/inventory/models.py
@@ -64,7 +64,7 @@ class ProductInventory(models.Model):
 
     quantity = models.IntegerField()
     restock_point = models.PositiveIntegerField(blank=True, null=True)
-    weight = models.IntegerField()  # in ounces
+    weight = models.FloatField()  # in ounces
     price = models.DecimalField(max_digits=10, decimal_places=2)
 
     class Meta:
